requirements.txt
numpy
scipy
scikit-learn
nltk
gensim
matplotlib
pandas
jsonlines
tweet-preprocessor


